requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
instagrapi==1.16.29
Pillow==9.2.0
pip==22.2.2
pycryptodomex==3.15.0
pydantic==1.9.2
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
setuptools==58.1.0
typing-extensions==4.3.0
urllib3==1.26.12
